phase2_knowledge_base/requirements.txt
aws-cdk-lib>=2.130.0
constructs>=10.3.0
boto3>=1.34.0
